+++ app/src/main/java/work/chiro/game/aircraft/MobEnemy.java
@@ -15,7 +15,7 @@ import java.util.LinkedList;
  * @author hitsz
  */
 public class MobEnemy extends AbstractAircraft {
-    public MobEnemy(AbstractConfig config, Vec2 posInit, AnimateContainer animateContainer, int hp) {
+    public MobEnemy(AbstractConfig config, Vec2 posInit, AnimateContainer animateContainer, double hp) {
         super(config, posInit, animateContainer, hp, 30);
     }
